Various render optimizations
This patch performs various rendering optimizations, giving a minor
performance improvement when using a Waveform source.

As the gradient shader never changes, it does not need to be reloaded
and recompiled every render frame, so this patch adds an instance member
to cache the compiled shader object. The shader is loaded and compiled
upon instance creation.

The vertex buffers used for curve and bars rendering are cached in a
similar way to avoid recreating and freeing them on every render frame.
Additionally, the curve vertex buffer is pre-populated upon creation to
avoid needing to update all three components of the vec3 when actually
rendering, instead only updating the y component.